Rename CTPUtils to ClientConfigurationUtils
